Lightweight extract allocation logic from mem_cache/common.py to more clearly show nearly parallel variants
